chore: move forking tests to a separate directory (#608)
